fix bug

+++ src/pages/Home.tsx
@@ -1,14 +1,19 @@
 
 import React, { useEffect, useState } from 'react';
 import { useQuery, useMutation, } from '@apollo/client'
-import { GET_CONTACT_LIST, DELETE_CONTACT, ADD_CONTACT } from '../queries';
-// import blob1 from "../assets/blob1.svg"
-// import blob2 from "../assets/blob2.svg"
+import Swal from 'sweetalert2';
 import useLocalStorage from '@/hooks/useLocalStorage';
+import blob1 from "@/assets/blob1.svg"
 import ListCard from '@/components/ListCard';
 import SearchForm from '@/components/SearchForm';
 import FavoriteCard from '@/components/FavoriteCard';
-import { Container, Card, Input, Button } from '@/components/ContactForm';
+import ButtonFloating from '@/components/Button';
+import { Container, Card, Input, Button, ButtonsContainer } from '@/components/ContactForm';
+import { Pagination, PaginationControlsContainer, LoadMoreButton, PageInfo, BackToPageButton, ContactsPerPageContainer } from '@/components/Pagination';
+import { GET_CONTACT_LIST } from '@/graphql/queries';
+import { ADD_CONTACT, DELETE_CONTACT } from '@/graphql/mutations';
+
+
 
 
 
@@ -25,7 +30,9 @@ interface ContactData {
 
 
 const Home: React.FC = () => {
-    const { loading, error, data } = useQuery<ContactData>(GET_CONTACT_LIST);
+    // State
+    const [limit, setLimit] = useState(10);
+    const [offset, setOffset] = useState(0);
     const [storedContacts, setStoredContacts] = useLocalStorage<Contact[]>('storedContacts', []);
     const [favoriteContacts, setFavoriteContacts] = useLocalStorage<Contact[]>('favoriteContacts', []);
     const [newContactData, setNewContactData] = useState({
@@ -37,40 +44,73 @@ const Home: React.FC = () => {
 
     });
     const [query, setQuery] = useState<string>("")
+    const [showForm, setShowForm] = useState(false);
+    const [formSubmitted, setFormSubmitted] = useState(false);
 
-
+    // GraphQL queries and mutations
+    const { loading, error, data } = useQuery<ContactData>(GET_CONTACT_LIST, {
+        variables: { limit, offset },
+    });
     const [deleteContact] = useMutation(DELETE_CONTACT);
     const [addContact] = useMutation(ADD_CONTACT);
 
+
+    // Effects
     useEffect(() => {
         if (data && data.contact) {
             setStoredContacts(data.contact);
         }
     }, [data, setStoredContacts]);
 
+    // Event Handlers
     const handleDeleteContact = async (id: number) => {
         try {
-            await deleteContact({ variables: { id } });
+            const result = await Swal.fire({
+                title: 'Delete Contact',
+                text: 'Are you sure you want to delete this contact?',
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#d33',
+                cancelButtonColor: '#3085d6',
+                confirmButtonText: 'Delete',
+                cancelButtonText: 'Cancel'
+            });
 
-            setStoredContacts(prevContacts => prevContacts.filter(contact => contact.id !== id));
+            if (result.isConfirmed) {
+                await deleteContact({ variables: { id } });
 
-            if (favoriteContacts.some(contact => contact.id === id)) {
-                const updatedFavoriteContacts = favoriteContacts.filter(contact => contact.id !== id);
-                setFavoriteContacts(updatedFavoriteContacts);
+                setStoredContacts(prevContacts => prevContacts.filter(contact => contact.id !== id));
+
+                if (favoriteContacts.some(contact => contact.id === id)) {
+                    const updatedFavoriteContacts = favoriteContacts.filter(contact => contact.id !== id);
+                    setFavoriteContacts(updatedFavoriteContacts);
+                }
+
+                Swal.fire({
+                    icon: 'success',
+                    title: 'Deleted!',
+                    text: 'The contact has been deleted.',
+                    confirmButtonText: 'OK'
+                });
             }
         } catch (error) {
             console.error('Error deleting contact:', error);
+            Swal.fire({
+                icon: 'error',
+                title: 'Error',
+                text: 'Error deleting contact: ' + error
+            });
         }
     };
-    console.log(storedContacts, favoriteContacts, "data old")
-    console.log(favoriteContacts, "datanew")
 
+    const fetchMoreContacts = () => {
+        setOffset(prevOffset => prevOffset + limit);
+    };
 
-
-
-    if (loading) return <p>Loading...</p>;
-    if (error) return <p>Error: {error.message}</p>;
-
+    const updateLimit = (newLimit: number) => {
+        setLimit(newLimit);
+        setOffset(0);
+    };
 
 
     const handleToggleFavorite = (id: number) => {
@@ -88,75 +128,11 @@ const Home: React.FC = () => {
         }
     };
 
+    const handleClick = () => {
+        setShowForm(prevShowForm => !prevShowForm);
+        setFormSubmitted(false);
+    };
 
-    //new contact
-    // const handleAddContact = async () => {
-    //     try {
-    //         const { firstName, lastName, phoneNumber } = newContactData;
-
-    //         // Check if the contact name is unique and doesn't have special characters
-    //         // const isNameValid = storedContacts.every(
-    //         //     contact =>
-    //         //         contact.first_name.toLowerCase() !== firstName.toLowerCase() &&
-    //         //         /^[a-zA-Z\s]*$/.test(firstName) && /^[a-zA-Z\s]*$/.test(lastName)
-    //         // );
-
-    //         // if (!isNameValid) {
-    //         //     console.error('Contact name must be unique and not contain special characters.');
-    //         //     return;
-    //         // }
-    //         const response = await addContact({
-    //             variables: {
-    //                 first_name: firstName,
-    //                 last_name: lastName,
-    //                 phones: [{ number: phoneNumber }],
-    //             },
-    //             update: (cache, { data }) => {
-    //                 if (data && data.insert_contact) {
-    //                     const newContact = data.insert_contact.returning[0];
-
-    //                     // Update the cache to include the new contact
-    //                     cache.modify({
-    //                         fields: {
-    //                             contact(existingContacts = []) {
-    //                                 const newContactRef = cache.writeFragment({
-    //                                     data: newContact,
-    //                                     fragment: gql`
-    //                                         fragment NewContact on Contact {
-    //                                             id
-    //                                             first_name
-    //                                             last_name
-    //                                             phones {
-    //                                                 number
-    //                                             }
-    //                                         }
-    //                                     `,
-    //                                 });
-
-    //                                 return [...existingContacts, newContactRef];
-    //                             },
-    //                         },
-    //                     });
-
-    //                     // Update local state and storage
-    //                     setStoredContacts(prevContacts => [...prevContacts, newContact]);
-    //                     localStorage.setItem('contactList', JSON.stringify([...storedContacts, newContact]));
-    //                 }
-    //             },
-    //         });
-
-    //         console.log('New contact added:', response);
-
-    //         // Reset form fields
-    //         setNewContactData({
-    //             firstName: '',
-    //             lastName: '',
-    //             phoneNumber: '',
-    //         });
-    //     } catch (error) {
-    //         console.error('Error adding new contact:', error);
-    //     }
-    // };
     const handleAddContact = async () => {
         try {
             const { firstName, lastName, phoneNumber, phoneNumber1, phoneNumber2 } = newContactData;
@@ -169,14 +145,15 @@ const Home: React.FC = () => {
             );
 
             if (!isNameValid) {
-                console.error('Contact name must be unique and not contain special characters.');
+                Swal.fire({
+                    icon: 'error',
+                    title: 'Oops...',
+                    text: 'Contact name must be unique and not contain special characters.'
+                });
                 return;
             }
 
-            if (phoneNumber === phoneNumber1 || phoneNumber === phoneNumber2 || phoneNumber1 === phoneNumber2) {
-                console.error('Phone numbers must be unique.');
-                return;
-            }
+
 
             const response = await addContact({
                 variables: {
@@ -204,18 +181,35 @@ const Home: React.FC = () => {
                     phoneNumber1: '',
                     phoneNumber2: ''
                 });
+                Swal.fire({
+                    icon: 'success',
+                    title: 'Contact Added',
+                    text: 'New contact has been successfully added!',
+                    confirmButtonText: 'OK'
+                });
+                setFormSubmitted(true);
             }
         } catch (error) {
-            console.error('Error adding new contact:', error);
+
+
+
+            Swal.fire({
+                icon: 'error',
+                title: 'Error',
+                text: 'Error adding new contact: ' + error
+            });
         }
     };
+
+    // Filtering
     const filteredContacts = storedContacts.filter(contact =>
         contact.first_name.toLowerCase().includes(query.toLowerCase()) ||
         contact.last_name.toLowerCase().includes(query.toLowerCase())
     );
 
-
-    console.log(filteredContacts, "data pencarian")
+    // Rendering
+    if (loading) return <p>Loading...</p>;
+    if (error) return <p>Error: {error.message}</p>;
 
     return (
         <>
@@ -223,49 +217,86 @@ const Home: React.FC = () => {
 
             <SearchForm setQuery={setQuery} query={query} />
 
-            {/* <img src={blob1} alt="blob1" className="cirlce1" />
-                <img src={blob2} alt="blob2" className="cirlce2" /> */}
+            <img src={blob1} alt="blob1" className="cirlce1" />
 
-            <Container>
+            <ButtonFloating onClick={handleClick} />
+            {showForm && !formSubmitted ? (<Container>
                 <Card>
-                    <label>First Name: </label>
+
                     <Input
                         type="text"
+                        placeholder="Input Your First Name.."
                         value={newContactData.firstName}
                         onChange={(e) => setNewContactData({ ...newContactData, firstName: e.target.value })}
                     />
-                    <label>Last Name: </label>
+
                     <Input
                         type="text"
+                        placeholder="Input Your Last Name.."
                         value={newContactData.lastName}
                         onChange={(e) => setNewContactData({ ...newContactData, lastName: e.target.value })}
                     />
-                    <div>
-                        <label>Phone Number: </label>
-                        <Input
-                            type="text"
-                            value={newContactData.phoneNumber}
-                            onChange={(e) => setNewContactData({ ...newContactData, phoneNumber: e.target.value })}
-                        />
-                    </div>
+
 
                     <Input
                         type="text"
-                        placeholder="Phone Number 2"
+                        placeholder="Input Your Phone Number.."
+                        value={newContactData.phoneNumber}
+                        onChange={(e) => setNewContactData({ ...newContactData, phoneNumber: e.target.value })}
+                    />
+
+
+
+                    <Input
+                        type="text"
+                        placeholder="Input Your Phone Number 2.."
                         value={newContactData.phoneNumber1}
                         onChange={e => setNewContactData({ ...newContactData, phoneNumber1: e.target.value })}
                     />
+
+
                     <Input
                         type="text"
-                        placeholder="Phone Number 3"
+                        placeholder="Input Your Phone Number 3.."
                         value={newContactData.phoneNumber2}
                         onChange={e => setNewContactData({ ...newContactData, phoneNumber2: e.target.value })}
                     />
-                    <Button onClick={handleAddContact}>Add Contact</Button>
+                    <ButtonsContainer>
+                        <Button onClick={handleAddContact}>Add Contact</Button>
+                        <Button onClick={handleClick}>Cancel</Button>
+                    </ButtonsContainer>
                 </Card>
 
-            </Container>
+            </Container>) : <>
 
+                {favoriteContacts.length > 0 && (
+                    <FavoriteCard favoriteContacts={favoriteContacts} onDelete={handleDeleteContact} onToggleFavorite={handleToggleFavorite} />
+                )}
+
+
+                <ListCard filteredContacts={filteredContacts} onDelete={handleDeleteContact} onToggleFavorite={handleToggleFavorite} />
+
+                <Pagination>
+                    <PaginationControlsContainer>
+                        <LoadMoreButton onClick={fetchMoreContacts}>Load More</LoadMoreButton>
+                        <PageInfo>
+                            Showing {offset + 1} - {Math.min(offset + limit, filteredContacts.length)} contacts
+                        </PageInfo>
+                        <ContactsPerPageContainer>
+                            <label>Contacts per page: </label>
+                            <select value={limit} onChange={e => updateLimit(Number(e.target.value))}>
+                                <option value={10}>10</option>
+                                <option value={20}>20</option>
+                                {/* Add more options as needed */}
+                            </select>
+                            <BackToPageButton onClick={() => setOffset(0)}>Back to Page 1</BackToPageButton>
+                        </ContactsPerPageContainer>
+                    </PaginationControlsContainer>
+                </Pagination>
+
+            </>}
+
+            {/* 
             <h1>Favorite</h1>
             {favoriteContacts.length > 0 && (
 
@@ -274,7 +305,7 @@ const Home: React.FC = () => {
             )}
 
             <h1 style={{ color: "white" }}>contact list</h1>
-            <ListCard filteredContacts={filteredContacts} onDelete={handleDeleteContact} onToggleFavorite={handleToggleFavorite} />
+            <ListCard filteredContacts={filteredContacts} onDelete={handleDeleteContact} onToggleFavorite={handleToggleFavorite} /> */}
         </>
     );
 };
